Agents/basic-strands-agents/lore-agent-basic/requirements.txt
strands-agents
strands-agents-tools
uv
uvicorn
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit